requirements.txt
altair==4.2.0
#attrs==22.1.0
attrs>=22.1.0,<23.0.0
blinker==1.5
blis==0.7.8
cachetools==5.2.0
#catalogue==1.0.0
catalogue>=2.0.6,<2.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
docx2txt==0.8
entrypoints==0.4
fonttools==4.37.2
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.3
murmurhash==1.0.8
nltk==3.7
numpy==1.23.2
packaging==21.3
pafy==0.5.5
pandas~=2.2.0
pathy==0.6.2
pdfminer.six==20220524
pdfminer3==2018.12.3.0
Pillow~=10.2.0
plac==1.1.3
plotly==5.10.0
preshed==3.0.7
protobuf==3.20.1
pyarrow==9.0.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
#pydeck==0.8.0b1
pydeck>=0.8.0b4,<1.0.0 
Pygments==2.13.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyresparser==1.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
regex==2022.8.17
requests==2.28.1
rich==12.5.1
semver==2.13.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sortedcontainers==2.4.0
#spacy==2.3.5
spacy==3.7.5
#spacy-legacy==3.0.11
spacy-legacy>=3.0.11,<3.1.0
spacy-loggers==1.0.3
#srsly==1.0.5
srsly==2.4.5
streamlit~=1.40.1
streamlit-tags==1.2.8
tenacity>=8.1.0,<10.0.0
#thinc==7.4.5
thinc>=8.2.2,<8.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.12.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
zipp==3.8.1
